# By Jonathan Chiam (2) and Pan Peining (2)
# Via Jonathan Chiam (1) and Pan Peining (1)
* 'PQL' of https://github.com/leonchinzeye/CS3201:
  Fixed ParentStarClause implementation and test bugs
  Fixed warning in build
  Add ParentsStarClauseTest
  Fix parent clause

+++ SPA/PDR.cpp
@@ -221,7 +221,7 @@ TNode* PDR::breakDownAssignExpression(ParsedData data, set<string>& usesSet) {
         }
     }
     
-    for(size_t i = 0; i < numExp; i++) {
+    for(int i = 0; i < numExp; i++) {
         string exp = expression.front();
         expression.pop();
